backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "The backend for the Onboard-me project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.14",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "supabase"
]
